+++ CHANGELOG.md
@@ -1,10 +1,10 @@
 ## 0.13.0 (2018-??-??)
 
-Adds the ability to run a `dry_run` of your benchmarks if you want to make sure
+Adds the ability to run a `pre_check` of your benchmarks if you want to make sure
 everything will run without error before running the full set of benchmarks.
 
 ### Features (User Facing)
-* new `dry_run` configuration option which allows users to add a dry run of all
+* new `pre_check` configuration option which allows users to add a dry run of all
 benchmarks with each input before running the actual suite. This should save
 time while actually writing the code for your benchmarks.